food_delivery_streaming/local/requirements.txt
# python deps for assignment
pyspark==3.4.0
psycopg2-binary==2.9.9
PyYAML==6.0.1
pandas==2.1.4
pyarrow==14.0.1
